# anms-core/integration_test/openapi.json
{
  "openapi": "3.0.2",
  "info": {
      "title": "AMMOS-ANMS",
      "description": "AMMOS-ANMS Docs",
      "version": "0.1.0"
  },
  "paths": {
...
      "/ari": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Paged Ari",
              "operationId": "paged_ARI_ari_get",
              "parameters": [
                  {
                      "required": false,
                      "schema": {
                          "title": "Page",
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 1
                      },
                      "name": "page",
                      "in": "query"
                  },
                  {
                      "required": false,
                      "schema": {
                          "title": "Size",
                          "maximum": 100.0,
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 50
                      },
                      "name": "size",
                      "in": "query"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Page_ARI_"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/ari/name/{obj_name}": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Ari By Name",
              "operationId": "ari_by_name_ari_name__obj_name__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Name",
                          "type": "string"
                      },
                      "name": "obj_name",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/ARI"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/formal_objects": {
          "get": {
              "tags": [
                  "ARIs",
                  "FORMAL"
              ],
              "summary": "Paged Formal Object",
              "operationId": "paged_formal_object_formal_objects_get",
              "parameters": [
                  {
                      "required": false,
                      "schema": {
                          "title": "Page",
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 1
                      },
                      "name": "page",
                      "in": "query"
                  },
                  {
                      "required": false,
                      "schema": {
                          "title": "Size",
                          "maximum": 100.0,
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 50
                      },
                      "name": "size",
                      "in": "query"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Page_FormalObject_"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/formal_objects/id/{obj_metadata_id}": {
          "get": {
              "tags": [
                  "ARIs",
                  "FORMAL"
              ],
              "summary": "Formal Object By Id",
              "operationId": "formal_object_by_id_formal_objects_id__obj_metadata_id__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Metadata Id",
                          "type": "integer"
                      },
                      "name": "obj_metadata_id",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/FormalObject"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/formal_objects/name/{obj_name}": {
          "get": {
              "tags": [
                  "ARIs",
                  "FORMAL"
              ],
              "summary": "Formal Object By Name",
              "operationId": "formal_object_by_name_formal_objects_name__obj_name__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Name",
                          "type": "string"
                      },
                      "name": "obj_name",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/FormalObject"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/formal_objects/edd/id/{obj_metadata_id}": {
          "get": {
              "tags": [
                  "ARIs",
                  "FORMAL"
              ],
              "summary": "Edd Formal By Id",
              "operationId": "edd_formal_by_id_formal_objects_edd_id__obj_metadata_id__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Metadata Id",
                          "type": "integer"
                      },
                      "name": "obj_metadata_id",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/EddFormal"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/formal_objects/mac/id/{obj_metadata_id}": {
          "get": {
              "tags": [
                  "ARIs",
                  "FORMAL"
              ],
              "summary": "Mac Formal By Id",
              "operationId": "mac_formal_by_id_formal_objects_mac_id__obj_metadata_id__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Metadata Id",
                          "type": "integer"
                      },
                      "name": "obj_metadata_id",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/MacFormal"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/formal_objects/rptt/id/{obj_metadata_id}": {
          "get": {
              "tags": [
                  "ARIs",
                  "FORMAL"
              ],
              "summary": "Rpt Formal By Id",
              "operationId": "rpt_formal_by_id_formal_objects_rptt_id__obj_metadata_id__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Metadata Id",
                          "type": "integer"
                      },
                      "name": "obj_metadata_id",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/RptFormal"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/formal_objects/ctrl/id/{obj_metadata_id}": {
          "get": {
              "tags": [
                  "ARIs",
                  "FORMAL"
              ],
              "summary": "Control By Id",
              "operationId": "control_by_id_formal_objects_ctrl_id__obj_metadata_id__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Metadata Id",
                          "type": "integer"
                      },
                      "name": "obj_metadata_id",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Control"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/formal_objects/ctrl/name/id/{obj_metadata_id}": {
          "get": {
              "tags": [
                  "ARIs",
                  "FORMAL"
              ],
              "summary": "Name Id Control",
              "operationId": "name_id_control_formal_objects_ctrl_name_id__obj_metadata_id__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Metadata Id",
                          "type": "integer"
                      },
                      "name": "obj_metadata_id",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "title": "Response Name Id Control Formal Objects Ctrl Name Id  Obj Metadata Id  Get",
                                  "type": "array",
                                  "items": {
                                      "$ref": "#/components/schemas/ControlNameId"
                                  }
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
  },
  "components": {
      "schemas": {
          "ARI": {
              "title": "ARI",
              "type": "object",
              "properties": {
                  "obj_metadata_id": {
                      "title": "Obj Metadata Id",
                      "type": "integer"
                  },
                  "obj_name": {
                      "title": "Obj Name",
                      "type": "string"
                  },
                  "adm_name": {
                      "title": "Adm Name",
                      "type": "string"
                  },
                  "type_name": {
                      "title": "Type Name",
                      "type": "string"
                  },
                  "namespace_id": {
                      "title": "Namespace Id",
                      "type": "integer"
                  },
                  "obj_id": {
                      "title": "Obj Id",
                      "type": "integer"
                  },
                  "parm_id": {
                      "title": "Parm Id",
                      "type": "integer"
                  },
                  "actual": {
                      "title": "Actual",
                      "type": "boolean"
                  },
                  "display": {
                      "title": "Display",
                      "type": "string"
                  }
              }
          },
...
          "Control": {
              "title": "Control",
              "type": "object",
              "properties": {
                  "obj_name": {
                      "title": "Obj Name",
                      "type": "string"
                  },
                  "namespace_id": {
                      "title": "Namespace Id",
                      "type": "integer"
                  },
                  "adm_name": {
                      "title": "Adm Name",
                      "type": "string"
                  },
                  "obj_formal_definition_id": {
                      "title": "Obj Formal Definition Id",
                      "type": "integer"
                  },
                  "fp_spec_id": {
                      "title": "Fp Spec Id",
                      "type": "integer"
                  },
                  "num_parms": {
                      "title": "Num Parms",
                      "type": "integer"
                  },
                  "parm_types": {
                      "title": "Parm Types",
                      "type": "string"
                  },
                  "parm_names": {
                      "title": "Parm Names",
                      "type": "string"
                  },
                  "parm_defaults": {
                      "title": "Parm Defaults",
                      "type": "string"
                  },
                  "use_desc": {
                      "title": "Use Desc",
                      "type": "string"
                  },
                  "obj_metadata_id": {
                      "title": "Obj Metadata Id",
                      "type": "integer"
                  }
              }
          },
          "ControlNameId": {
              "title": "ControlNameId",
              "type": "object",
              "properties": {
                  "obj_name": {
                      "title": "Obj Name",
                      "type": "string"
                  },
                  "obj_metadata_id": {
                      "title": "Obj Metadata Id",
                      "type": "integer"
                  },
                  "num_parms": {
                      "title": "Num Parms",
                      "type": "integer"
                  }
              }
          },
          "EddFormal": {
              "title": "EddFormal",
              "type": "object",
              "properties": {
                  "obj_metadata_id": {
                      "title": "Obj Metadata Id",
                      "type": "integer"
                  },
                  "obj_name": {
                      "title": "Obj Name",
                      "type": "string"
                  },
                  "namespace_id": {
                      "title": "Namespace Id",
                      "type": "integer"
                  },
                  "obj_formal_definition_id": {
                      "title": "Obj Formal Definition Id",
                      "type": "integer"
                  },
                  "fp_spec_id": {
                      "title": "Fp Spec Id",
                      "type": "integer"
                  },
                  "use_desc": {
                      "title": "Use Desc",
                      "type": "string"
                  }
              }
          },
          "FormalObject": {
              "title": "FormalObject",
              "type": "object",
              "properties": {
                  "obj_metadata_id": {
                      "title": "Obj Metadata Id",
                      "type": "integer"
                  },
                  "type_name": {
                      "title": "Type Name",
                      "type": "string"
                  },
                  "obj_name": {
                      "title": "Obj Name",
                      "type": "string"
                  },
                  "namespace_id": {
                      "title": "Namespace Id",
                      "type": "integer"
                  },
                  "adm_name": {
                      "title": "Adm Name",
                      "type": "string"
                  },
                  "adm_enum": {
                      "title": "Adm Enum",
                      "type": "integer"
                  },
                  "adm_enum_label": {
                      "title": "Adm Enum Label",
                      "type": "integer"
                  },
                  "use_desc": {
                      "title": "Use Desc",
                      "type": "string"
                  },
                  "obj_formal_definition_id": {
                      "title": "Obj Formal Definition Id",
                      "type": "integer"
                  },
                  "formal_desc": {
                      "title": "Formal Desc",
                      "type": "string"
                  },
                  "is_formal": {
                      "title": "Is Formal",
                      "type": "boolean",
                      "default": false
                  }
              }
          },
...
          "HTTPValidationError": {
              "title": "HTTPValidationError",
              "type": "object",
              "properties": {
                  "detail": {
                      "title": "Detail",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/ValidationError"
                      }
                  }
              }
          },
...
          "MacFormal": {
              "title": "MacFormal",
              "type": "object",
              "properties": {
                  "obj_metadata_id": {
                      "title": "Obj Metadata Id",
                      "type": "integer"
                  },
                  "obj_name": {
                      "title": "Obj Name",
                      "type": "string"
                  },
                  "namespace_id": {
                      "title": "Namespace Id",
                      "type": "integer"
                  },
                  "obj_formal_definition_id": {
                      "title": "Obj Formal Definition Id",
                      "type": "integer"
                  },
                  "fp_spec_id": {
                      "title": "Fp Spec Id",
                      "type": "integer"
                  },
                  "use_desc": {
                      "title": "Use Desc",
                      "type": "string"
                  }
              }
          },
          "Page_ARI_": {
              "title": "Page[ARI]",
              "required": [
                  "items",
                  "total",
                  "page",
                  "size"
              ],
              "type": "object",
              "properties": {
                  "items": {
                      "title": "Items",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/ARI"
                      }
                  },
                  "total": {
                      "title": "Total",
                      "minimum": 0.0,
                      "type": "integer"
                  },
                  "page": {
                      "title": "Page",
                      "minimum": 1.0,
                      "type": "integer"
                  },
                  "size": {
                      "title": "Size",
                      "minimum": 1.0,
                      "type": "integer"
                  }
              }
          },
...
          "Page_FormalObject_": {
              "title": "Page[FormalObject]",
              "required": [
                  "items",
                  "total",
                  "page",
                  "size"
              ],
              "type": "object",
              "properties": {
                  "items": {
                      "title": "Items",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/FormalObject"
                      }
                  },
                  "total": {
                      "title": "Total",
                      "minimum": 0.0,
                      "type": "integer"
                  },
                  "page": {
                      "title": "Page",
                      "minimum": 1.0,
                      "type": "integer"
                  },
                  "size": {
                      "title": "Size",
                      "minimum": 1.0,
                      "type": "integer"
                  }
              }
          },
...
          "RptFormal": {
              "title": "RptFormal",
              "type": "object",
              "properties": {
                  "obj_metadata_id": {
                      "title": "Obj Metadata Id",
                      "type": "integer"
                  },
                  "obj_name": {
                      "title": "Obj Name",
                      "type": "string"
                  },
                  "namespace_id": {
                      "title": "Namespace Id",
                      "type": "integer"
                  },
                  "obj_formal_definition_id": {
                      "title": "Obj Formal Definition Id",
                      "type": "integer"
                  },
                  "fp_spec_id": {
                      "title": "Fp Spec Id",
                      "type": "integer"
                  },
                  "use_desc": {
                      "title": "Use Desc",
                      "type": "string"
                  }
              }
          },
...
          "ValidationError": {
              "title": "ValidationError",
              "required": [
                  "loc",
                  "msg",
                  "type"
              ],
              "type": "object",
              "properties": {
                  "loc": {
                      "title": "Location",
                      "type": "array",
                      "items": {
                          "anyOf": [
                              {
                                  "type": "string"
                              },
                              {
                                  "type": "integer"
                              }
                          ]
                      }
                  },
                  "msg": {
                      "title": "Message",
                      "type": "string"
                  },
                  "type": {
                      "title": "Error Type",
                      "type": "string"
                  }
              }
          },
...
      }
  }
}